[염정운] Controller, Service, Dto, Repository를 지시한 대로 리펙터링
