--- src/main/java/com/hugudungs/hugupjigup/dto/comment/freecomment/FreeCommentUpdateResponseDto.java
@@ -1,17 +1,0 @@
-package com.hugudungs.hugupjigup.dto.comment.freecomment;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class FreeCommentUpdateResponseDto {
-    private Long id;
-    private String content;
-    private Long userId;
-    private Long freeId;
-}
